Updated test.py, renamed tests dir
Added option to show a test/vimrc description
Also an option to use different default vimrc's using -v argument
