+++ vimrunner/default/alt.vim
@@ -1,0 +1,11 @@
+" DESCRIPTION: alt vimrc (nocompatible)
+" needed by vimrunner
+function! VimrunnerPyEvaluateCommandOutput(command)
+  redir => output
+    silent exe a:command
+  redir END
+  return output
+endfunction
+
+set runtimepath^=..
+set nocompatible
